+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.9.2
+- MiniMap improviments. Fix issue [#517](https://github.com/RafaelBarbosatec/bonfire/issues/517)
+- Raname `BouncingObject` to `ElasticCollision`.
+- Fix `SimpleDirectionAnimation` bug when render fastAnimation.
+
 # 3.9.1
 - `BlockMovementCollision` improvements.
 - Create a `PinchGesture` mixin to listen pinch gestures.